commit message: refactor(model): standardize data types in `ModelArgs` struct
commit description:
- Updated integer fields in `ModelArgs` from `int` to `int64_t` for consistency and support of larger values.
- Replaced `float` types with `double` for `norm_eps` and `dropout` to maintain precision in model configuration.
- Ensured adherence to standardized data types for improved compatibility and numerical stability.

This refactor improves the robustness and adaptability of the `ModelArgs` struct, aligning it with best practices for data type usage in model configuration.

+++ src/model/ModelArgs.h
@@ -5,16 +5,16 @@
 
 namespace Model {
     struct ModelArgs {
-        int dim = 4096;
-        int n_layers = 32;
-        int n_heads = 32;
-        std::optional<int> n_kv_heads = std::nullopt;
-        int vocab_size = 32000;
-        std::optional<int> hidden_dim = std::nullopt;
-        int multiple_of = 256;
-        float norm_eps = 1e-5;
-        int max_seq_len = 2048;
-        float dropout = 0.0;
+        int64_t dim = 4096;
+        int64_t n_layers = 32;
+        int64_t n_heads = 32;
+        std::optional<int64_t> n_kv_heads = std::nullopt;
+        int64_t vocab_size = 32000;
+        std::optional<int64_t> hidden_dim = std::nullopt;
+        int64_t multiple_of = 256;
+        double norm_eps = 1e-5;
+        int64_t max_seq_len = 2048;
+        double dropout = 0.0;
     };
 }
